django/requirements.txt
appdirs==1.4.4
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
Django==4.1
djangorestframework==3.14.0
frozendict==2.3.5
html5lib==1.1
idna==3.4
lxml==4.9.2
multitasking==0.0.11
numpy==1.24.2
pandas==1.5.3
psycopg2==2.9.5
pycoingecko==3.1.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.14
webencodings==0.5.1
yfinance==0.2.12